requirements.txt
# Flask Database Starter - Requirements
# Install all: pip install -r requirements.txt

# Core
flask>=2.0.0

# Database ORM
flask-sqlalchemy>=3.0.0

# Migrations
flask-migrate>=4.0.0

# Authentication
flask-login>=0.6.0
werkzeug>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# PostgreSQL driver (uncomment if needed)
# psycopg2-binary>=2.9.0

# MySQL driver (uncomment if needed)
# pymysql>=1.0.0
